Fix compiler warnings about narrowing conversions and unused variables and parameters.
(cherry picked from commit 5fbce61aa16014cd14d96905cac820082dc49a59)

+++ test/algorithm/for_test.cpp
@@ -34,8 +34,8 @@ namespace
 
         std::string test_str("(123456789)");
         std::ostringstream out;
-        int iii;
-        int size = v.size();
+        size_t iii;
+        size_t size = v.size();
         (
             out << val("("),
             for_(ref(iii) = 0, ref(iii) < size, ++ref(iii) )
